refactor: make proper module separation
